requirements.txt
beautifulsoup4
pandas
Pillow
pydantic
pyopenssl
requests>=2.25.1
XlsxWriter